change main.js to main.ts
