+++ package.json
@@ -2,7 +2,7 @@
   "name": "quiz-app",
   "version": "0.0.0",
   "license": "MIT",
-  "main": "main.js",
+  "main": "main.ts",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
